% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{randomly_assign}
\alias{randomly_assign}
\title{Assigns n samples into k groups}
\usage{
randomly_assign(n, k)
}
\arguments{
\item{n}{sample size}

\item{k}{number of groups}
}
\value{
returns a vector of length n with a random assignment of entries from 1 to k
}
\description{
Assigns n samples into k groups
}
\examples{
n <- 100
folds_10 <- randomly_assign(n,10)
folds_5 <- randomly_assign(n,5)
}
\author{
Ben Sherwood <ben.sherwood@ku.edu>
}

