% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roccv.r
\docType{package}
\name{roccv}
\alias{roccv}
\alias{roccv-package}
\title{roccv: A package for creating ROC plots on cross validated data}
\description{
roccv: A package for creating ROC plots on cross validated data
}

